Update main.py
Added whatsapp_login function. Now the whatsapp_checker function can check a list of phone numbers instead of one, which greatly improves the speed of the application. And as a bonus, the execution result is saved in a csv file so that you don't have to run the program to check the same phone.